+++ packages/MSBot/src/msbot.ts
@@ -48,7 +48,7 @@ program
     .command('export', 'export all connected services so that this bot can be cloned');
 
 program
-    .command('clone', 'create a new .bot file and all of its resources based on an exported .recipe file and folder ');
+    .command('clone', 'create a new Azure Group and clone all of the .recipe services into it. ');
 
 program
     .command('connect <service>', 'connect to a resource (Luis/Qna/Azure/...) used by the bot');
